Refactoring graphql
